MUST_10K_LIMIT: 1000 --> 10_000

+++ tests/accuracy/test_acc_util/bdd100k.py
@@ -55,7 +55,7 @@ def to_numpy(det_boxes: ObjectDetectionResult) -> np.ndarray:
     for b in det_boxes:
         box = b.boundingbox
         d.append([box.left, box.top, box.right, box.bottom, b.score, b.index])
-    return np.array(d)
+    return np.array(d, dtype=np.float32)
 
 
 class MAPMetricYolov5:
